validation styling. changed routes page to dashboard and added a routes page.
