+++ src/app/page.tsx
@@ -1,9 +1,11 @@
 'use client';
 
+import Image from 'next/image';
+import { useAuth } from '@/contexts/auth-context-provider';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useForm } from 'react-hook-form';
+import { useForm, useFormState } from 'react-hook-form';
 import { z } from 'zod';
 import bcrypt from 'bcryptjs';
 import BusPhotoOne from '@/public/bus-one.jpg';
@@ -21,8 +23,6 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import Image from 'next/image';
-import { useAuth } from '@/contexts/auth-context-provider';
 
 const registerFormSchema = z
   .object({
@@ -49,6 +49,8 @@ const RegisterForm = () => {
     },
   });
 
+  const { errors } = useFormState({ control: form.control });
+
   async function onSubmit(data: z.infer<typeof registerFormSchema>) {
     const users = JSON.parse(localStorage.getItem('users') || '[]');
     const hashedPassword = await bcrypt.hash(data.password, 12);
@@ -65,13 +67,13 @@ const RegisterForm = () => {
   }
 
   if (isLoggedIn) {
-    router.push('/routes');
+    router.push('/dashboard');
   }
 
   return (
     <div className="relative flex justify-center w-full h-screen">
-      <div className="absolute md:grid md:grid-cols-2 grid-rows-2 w-full h-full hidden">
-        <div className="relative w-full h-full">
+      <div className="absolute md:grid md:grid-cols-2 grid-rows-2 w-full h-full">
+        <div className="relative w-full h-full hidden md:block">
           <Image
             src={BusPhotoOne}
             alt="Bus Photo One"
@@ -81,7 +83,7 @@ const RegisterForm = () => {
             priority
           />
         </div>
-        <div className="relative w-full h-full">
+        <div className="relative w-full h-full hidden md:block">
           <Image
             src={BusPhotoTwo}
             alt="Bus Photo Two"
@@ -101,7 +103,7 @@ const RegisterForm = () => {
             priority
           />
         </div>
-        <div className="relative w-full h-full">
+        <div className="relative w-full h-full hidden md:block">
           <Image
             src={BusPhotoThree}
             alt="Bus Photo Four"
@@ -130,7 +132,11 @@ const RegisterForm = () => {
                   name="name"
                   render={({ field }) => (
                     <FormItem>
-                      <FormLabel>Name:</FormLabel>
+                      <FormLabel
+                        className={`${errors.name ? 'text-red-700' : ''}`}
+                      >
+                        Name:
+                      </FormLabel>
                       <FormControl>
                         <Input
                           className="w-full"
@@ -139,7 +145,7 @@ const RegisterForm = () => {
                           {...field}
                         />
                       </FormControl>
-                      <FormMessage />
+                      <FormMessage className="text-red-700" />
                     </FormItem>
                   )}
                 />
@@ -148,7 +154,11 @@ const RegisterForm = () => {
                   name="email"
                   render={({ field }) => (
                     <FormItem>
-                      <FormLabel>Email:</FormLabel>
+                      <FormLabel
+                        className={`${errors.email ? 'text-red-700' : ''}`}
+                      >
+                        Email:
+                      </FormLabel>
                       <FormControl>
                         <Input
                           className="w-full"
@@ -157,7 +167,7 @@ const RegisterForm = () => {
                           {...field}
                         />
                       </FormControl>
-                      <FormMessage />
+                      <FormMessage className="text-red-700" />
                     </FormItem>
                   )}
                 />
@@ -166,7 +176,11 @@ const RegisterForm = () => {
                   name="password"
                   render={({ field }) => (
                     <FormItem>
-                      <FormLabel>Password:</FormLabel>
+                      <FormLabel
+                        className={`${errors.password ? 'text-red-700' : ''}`}
+                      >
+                        Password:
+                      </FormLabel>
                       <FormControl>
                         <Input
                           className="w-full"
@@ -175,7 +189,7 @@ const RegisterForm = () => {
                           {...field}
                         />
                       </FormControl>
-                      <FormMessage />
+                      <FormMessage className="text-red-700" />
                     </FormItem>
                   )}
                 />
@@ -184,7 +198,13 @@ const RegisterForm = () => {
                   name="confirmPassword"
                   render={({ field }) => (
                     <FormItem>
-                      <FormLabel>Confirm Password:</FormLabel>
+                      <FormLabel
+                        className={`${
+                          errors.confirmPassword ? 'text-red-700' : ''
+                        }`}
+                      >
+                        Confirm Password:
+                      </FormLabel>
                       <FormControl>
                         <Input
                           className="w-full"
@@ -193,7 +213,7 @@ const RegisterForm = () => {
                           {...field}
                         />
                       </FormControl>
-                      <FormMessage />
+                      <FormMessage className="text-red-700" />
                     </FormItem>
                   )}
                 />
